refactor(be): 주문 도메인에서 현재 부적합한 데이터 타입 변경
* OrderItem 에서 추후 값 연산 시 값의 오버플로우 방지
* 현재 설계상 데이터 값을 소수점으로 만들어서 저장하지는 않으므로 BigDecimal 대신 Long 사용

+++ backend/src/main/java/com/deliveranything/domain/order/dto/OrderResponse.java
@@ -3,7 +3,6 @@ package com.deliveranything.domain.order.dto;
 import com.deliveranything.domain.order.entity.Order;
 import com.deliveranything.domain.order.entity.OrderItem;
 import com.deliveranything.domain.order.enums.OrderStatus;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,9 +15,9 @@ public record OrderResponse(
     String address,
     String riderNote,
     String storeNote,
-    BigDecimal totalPrice,
-    BigDecimal storePrice,
-    BigDecimal deliveryPrice,
+    Long totalPrice,
+    Long storePrice,
+    Long deliveryPrice,
     LocalDateTime createdAt
 ) {
